pull state randomization functions out of rewards testing
